bi-directional gateway.c
I have added a few lines of code which i currently have working in my system to set a relay output on a node from openHAB. 

line 88: assigned an additional payload structure to handle outgoing message
lines 139 - 140 (can be expanded): subscribe to topics (these should be in the form (AABC) where AA is the node number, B is the device/sensor number and C is the structure element
lines 299-320: this is where the payload is passed to the node. right now it is setup to handle a switch state change in the first structure variable

this is what my item in openHAB looks like:

Switch itm_gDoor_control  "Switch Garage Door" {mqtt=">[mymosquitto:5022:command:ON:1],>[mymosquitto:5022:OFF:0]"}

then i add the switch to my sitemap and i can control a relay on a node by subscribing to topic 5022 in the gateway. I also created a rule in openHAB to switch the switch state to OFF after it sees the state of the garage door change.